Create users API - notify users when already registered
